🚸 Fall back to raw sha when provided sha is invalid
The Percy API will respond with an error if the git sha is less than 40
characters. Most cases when this is an issue, the CI environment variable is
truncated and is fixed by setting it to the raw git sha. We can handle this more
seemlessly by providing this same fallback internally in @percy/env