🔧 Generate oclif manifests after building
The publish hooks are recommended when using TypeScript since, in development,
oclif will use TypeScript module resolutions. When building oclif without
TypeScript, oclif will use globby to resolve CLI commands when the manifest is
missing. By generating the manifest after building, using the CLI in development
is quicker and does not rely on globby.

The generation of manifests still looks at the filesystem. So dependent packages
need to be finished building before the manifest is generated. Using lerna
--stream rather than --parallel will respect package topology.