✨ Accept a directory path for autoloading configs
The `filepath` option was renamed to `path` and can now be a path to a directory
containing a Percy config file.

The test helper was updated to use strings rather than objects and mock fs rather
than cosmiconfig. As a result, the tests were updated to provide strings to the
mock helper rather than objects.