✨ Add --parallel exec flag (#51)
* 🔥 Remove auto-detect for parallel total

The presence of this attribute in the API will mark a build as parallel or not. In most cases, the
user sets this total to -1 and manually finalizes the build. In a few select CI services, this
parallel total is automatically detected based on the assumption that Percy itself is being
parallelized. However, this is not always true, and in a lot of cases Percy is only being run in one
of the parallel nodes. This causes the build to be stuck waiting for other, non-existent, parallel
builds to finalize.

Going forward, the easiest and quickest option will be to define the parallel total when, and only
when, you wish to parallelize Percy builds specifically. Finalize will be recommended, but
explicitly setting this to a known number or other env var will still enable the option to skip
manual finalization at the end of all parallel builds.

* 💥 Rename port char

This option is rarely used, and the existing char can be utilized elsewhere

* ✨ Mark builds as parallel with a --parallel flag

This flag will automatically set the env var, PERCY_PARALLEL_TOTAL, to "-1"

* ✨ Default parallel total when not defined