:sparkles: Add backwards compatability for DOM library fetching
This allows us to use the new CLI with any non-JS SDK right now. This is because
the only breaking changes for those non-js SDKs was:

- Change in server path for where to get the DOM library
- Change in function signature for how the dom library works

This commit adds back the `percy-agent.js` server endpoint (with a deprecation
log), and also adds a wrapper to match the old function signature.