✨ Even better logging (#140)
* ✨ Add internal class for launching and communicating with a browser

* ✨ Replace winston with custom logger

* ✨ Use new logger style for all packages

* ✨ Add option to print config logs without debug enabled

* ✅ Use new logger style throughout tests

* ✅ Create new logging test helper

* ✅ Use new logging test helper

* 🐛 Always log to the default logging streams

* 💥 Use new logger for sdk-utils

* ♻ Refactor capture method helpers

For logging purposes, these functions are better off not as helpers due to how specific they are to
a page and its assocated capture or snapshot information

* ✅ Fix flakey test

With additional logging, it became evident that the page was being closed during navigation. This is
handled by a navigation timeout, however that timeout is too long to account for in this test. The
test has been adjusted to help ensure navigation is finished when closing.

* 🐛 Allow errors to bubble from wait-for util

* 🐛 Immediately abort network idle if page closes

* 🐛 Abort navigation when page is closed

* ✅ Add test for closing during network idle

* ✅ Give test more time to run

* 🔥 Remove file left from previous rebasing

* 📝 Update @percy/logger readme