✨ Lazily create sdk-util logger for testing purposes (#145)
A logger is a singleton, but tests teardown and recreate that singleton for a clean
environment. When a logging group is created, it is bound to the singleton instance at the time of
creation. Creating a group at the root level causes it to forever be bound to that initial instance
and not be effected by test setup. By lazily creating the logging group, we can be assured that the
test environment will be cleaned up without requiring a re-require of the file where the logger is
created.