✅ Fix flakey tests (#184)
* ✅ Improve frame serialization test

Sometimes a frame may take longer than expected to load. Even the JS only frame may load after
modifying the frame document causing that test to flake. On windows, frames can take even longer to
load than 3s, so the wait timeout was upped to 5s and the hook timeout was removed.

* ✅ Delay test result slightly for reliability

It's possible for the test to continue after an asset has been accessed but before the request has
been intercepted causing it to flake. Delaying when asset access is detected should help the
request be intercepted in time for the assertion.

* ✅ Make test more deterministic

Sometimes it can take more than half of a second for the server to shutdown when forcefully
terminating. Removing the arbitrary wait and making the test more deterministic should show if this
was a test timing issue, or if termination itself is flakey.

* ✅ Add test to cover browser closed error messages