✅ Remove explicit browser install test
This script must work for any tests to work. Since it is technically tested by running before tests,
it does not generate coverage. In order to generate coverage, the existing browser install directory
is deleted within the test. This causes issues on Windows and actually makes the test suite slower
overall. Removing the test solves the issue on Windows, speeds up the test suite, and is still
required to work for almost any tests.