💚 Add Windows workflow (#176)
* 👷 Add Windows CI workflow

* 🔨 Fix Windows babel registration

* ✅ Fix cli-config tests for Windows

* ✅ Fix config tests for Windows

* 🔇 Ignore rimraf errors on tmp cleanup

* ⬆ Upgrade interactor.js from 2.0.0-beta.9 to 2.0.0-beta.10

* ✅ Remove explicit browser install test

This script must work for any tests to work. Since it is technically tested by running before tests,
it does not generate coverage. In order to generate coverage, the existing browser install directory
is deleted within the test. This causes issues on Windows and actually makes the test suite slower
overall. Removing the test solves the issue on Windows, speeds up the test suite, and is still
required to work for almost any tests.

* ✅ Use platform agnostic command for exec tests

* 🐛 End browser process pipes during cleanup

This is needed specifically for Windows