✨ Add ability to download browser during post-install (#206)
* ♻ Refactor core browser install script

This script is an integral part of core's asset discovery process and more than a simple utility
function. The base install function was made to be more agnostic of what file is being downloaded
and extracted. Chromium install defaults were moved to their own function rather than being defaults
of the agnostic install function.

* 🔨 Resolve monorepo packages to src files in development

Allows tests and scripts to run without always requiring a build step

* ✨ Add ability to download browser during post-install

* 💚 Optimize CI workflow caching

* ✅ Fix @percy/env test helper

The built file is no longer used and this line results in an error when the file does not exist

* 🔧 Add production env var to build scripts