♻ Refactor discovery queue
The queue has been refactored to accept ids for tasks and allow pausing & flushing. This allows us
to add a deferUploads option which, when enabled, will wait until the process is stopped before
creating a build and uploading all snapshots. While deferred, duplicate snapshots are replaced in
the queue which allows SDKs to retry snapshots without duplicate name errors from the upstream API.

The discoverer class has been converted to be functional helpers that the network class can use to
enable request interception. This allows the capture method to intercept resources while taking a
DOM snapshot. With this addition, the capture method no longer requires the snapshot method, however
the two methods then became so similar, that it made sense move the capture functionality to the
snapshot method entirely.

With the snapshot method now accepting capture options, a per-snapshot config util was created to
validate and merge all snapshot options. Part of this validation is ensuring that capture options
are not provided along with a static dom snapshot. The snapshots capture option was also renamed to
be additionalSnapshots to avoid confusion with the snapshot method.

With the major parts of this library shifting around, other changes were also made in
occordance. The src file tree was flattened, util methods were combined, and tests were
reorganized to match the new code structure.