♻ Refactor config utils
The underlying merge util was refactored to accept an array of objects instead of only two
objects. The options argument was replaced with a map callback argument. The function will be called
for each property within each object recursively with the path, previous value, and next value of
the property. A path-value pair can be returned to override the default merge behavior.

Some options previously handled by the old merge function were replaced by the new map callback in
the appropriate places.

The validate util was also refactored, though to a much less degree. Default schema $id's now start
with a forward slash to help disguish them from local schema references. The function will also now
only return an array of errors on failure and undefined on success. The associated addSchema util
was also updated to accept schema's without automatically adding them to the global Percy config
schema. This allows the validate util to be used with other registered schemas and not just the
global Percy config schema.