♻ Move logic fork for loading config files
The load method has the capability to not load files, but the "using config" log would log twice if
the load method was called twice. The solution for silencing the duplicate log was to not call load
a second time and instead just load defaults. This prompted a test for this new code path for
coverage at the time.

However, now that validation and normalization have changed, we can adjust this within the load
method itself. If a file is not loaded, the log will not happen; but validation will still be
performed within the load method in addition to merging with default as the old additional code path
did. This actually improves additional config file options by always validating, even when not
loaded by a config file.