✨ Normalize config with schemas (#450)
* ✅ Add test for normalization by schema

* 🔥 Remove now unnecessary extra normalization

Empty values as a result of pruning invalid options are now already removed

* 🐛 Fix merge util array mapping

When mapping parent keys of arrays the context is lost, which can trick the following logic into
thinking it is creating a new array rather than concatenating arrays.

* ♻ Export more config utils

* ✨ Allow fetching nested schemas via an accumulated array

If a path is provided, subsequent nested schemas will be accumulated with the original root schema
by recursively fetching the next path's schema. To facilitate schema references, the base
functionality of this method will now resolve any `$ref` in a provided schema by once again
recursively fetching the desired schema and plucking out the referenced meta path. For schemas that
actually contain more than a single possible schema, such as tuples or meta schemas like `oneOf`,
the best matching schema is determined by how deep the paths match; with a bias towards loosely
matched schemas that can contain properties not defined by said schema.

* ✨ Allow normalizing config keys based on their schema

This eliminates the need for a constant to determine which objects should not have their properties
normalized. While this constant is easy to update for internal packages, it can cause issues for
external packages that require certain properties not be normalized.

* 🐛 Handle non-strings in normalize casing functions

* ✨ Allow migrations for specific schemas

Now that a property's normalization can depend on its schema, the migrate function can now also accept
a schema name so it can properly normalize any incoming data.

Since normalization can be dependant on a schema, and because migrations normalize their data, the
migrate helpers were updated to allow defining migrations for specific schemas. While this won't be
utilized immediately for migrating other options, it can definitely come in handy for future
deprecations or breaking changes.

* ✨ Normalize loaded config files against the default config schema

* 🐛 Update corresponding schemas that should not be normalized

* ♻ Move snapshot options normalization to snapshot config helper

* ♻ Move logic fork for loading config files

The load method has the capability to not load files, but the "using config" log would log twice if
the load method was called twice. The solution for silencing the duplicate log was to not call load
a second time and instead just load defaults. This prompted a test for this new code path for
coverage at the time.

However, now that validation and normalization have changed, we can adjust this within the load
method itself. If a file is not loaded, the log will not happen; but validation will still be
performed within the load method in addition to merging with default as the old additional code path
did. This actually improves additional config file options by always validating, even when not
loaded by a config file.

* 🐛 Update normalize usage in cli-config

* ✅ Update test for normalized provided options

* ✅ Update test to not use empty values