✨ Allow registering multiple config migrations at a time (#456)
* ✨ Allow registering multiple migrations at a time

* 🐛 Do not force version in migrate function

Instead, a default config migration can be registered to accomplish the same.

Migrations are now registered in reverse order, a.k.a. last-in-first-out. This allows the default
migration to always run last.

* 📝 Update normalize doc comment

* ♻ Move deprecated snapshot config to own migrate function

* 🐛 Fix exec init hook migration registration

* 🔊 Warn on every snapshot deprecated migration

We want to show this warning anytime the migration detects them. A deprecation log will only show
once per message. This was the previous behaviour, but accidentally got dropped when it was moved