🐛 Track and properly cleanup frame requests
If a request for a document results in a new session, such as for isolated pages, the frame is
detached before the request's loading finished event gets a chance to fire. Requests for new frames
can be tracked and when the frame becomes detached the associated request can be cleaned up.