♻ Refactor page navigation to be more reliable
The existing lifecycle handling combined with the polling nature of navigation resolution could
sometimes result in a situation where a page's lifecycle is reset before navigation can be deemed
complete. We can use the same technique watching for the proper lifecycle as we do for frame
navigation. However, rather than repeating code within the handlers, they were refactored slightly
to be mapped handlers. The navigation trigger was also refactored slightly into a callback that can
be called the moment we start waiting for the navigation handlers to finish.