🐛 Fix flakey redirect handling
Do to certain race conditions, intecepted redirects were not properly getting associated with an
intercept id. But since the previous request in the chain had been cleaned up, this could sometimes
cause asset discovery to hang. Adjusting the if-else logic of this handling fixes this issue by
ensuring that if a request is not being handled, it is always set up to be intercepted.