🐛 Fix situations where asset discovery might hang (#490)
* 🔊 Log a better message for network idle timeouts

* ♻ Move page resizing out of page init

🐛 And always pass the minHeight when resizing

* 🐛 Disable browser translate feature correctly

* 🐛 Automatically initialize subframe sessions

This allows request interception bindings to propagate down to subframes

* ♻ Unprivate some page properties

* 🐛 Track and properly cleanup frame requests

If a request for a document results in a new session, such as for isolated pages, the frame is
detached before the request's loading finished event gets a chance to fire. Requests for new frames
can be tracked and when the frame becomes detached the associated request can be cleaned up.

* ✅ Prefer localhost subdomains for some tests

* 🙈 Adjust coverage comments

* ♻ Refactor page navigation to be more reliable

The existing lifecycle handling combined with the polling nature of navigation resolution could
sometimes result in a situation where a page's lifecycle is reset before navigation can be deemed
complete. We can use the same technique watching for the proper lifecycle as we do for frame
navigation. However, rather than repeating code within the handlers, they were refactored slightly
to be mapped handlers. The navigation trigger was also refactored slightly into a callback that can
be called the moment we start waiting for the navigation handlers to finish.

* 🐛 Fix flakey redirect handling

Do to certain race conditions, intecepted redirects were not properly getting associated with an
intercept id. But since the previous request in the chain had been cleaned up, this could sometimes
cause asset discovery to hang. Adjusting the if-else logic of this handling fixes this issue by
ensuring that if a request is not being handled, it is always set up to be intercepted.