🚨 Lint against extraneous dependencies
In all circumstances where it matters in this repo, the extraneous dependencies are guaranteed to be
present and resolve to their expected versions. However, to better support Yarn 2+ and any other
potential node loaders that might block on this, this eslint rule will help ensure that packages
only import from their specific dependencies.