✨ Serialize current video frame into a `poster` on the `<video>` (#794)
* ✨ Serialize current video frame into a `poster` on the `<video>`

Capturing videos via request interception can be tricky and highly depends on
the customers server. Percy's infrastructure will also not play videos for
screenshots so it doesn't make too much sense to capture large resources like
vidoes.

This feature will capture the current frame of the videos on page and turn them
into a poster image for the video. This is so there's something to display on
the `<video>` element when capturing a screenshot in Percy's browsers.

It's possible for this to introduce flakes since this will capture the current
frame of the video. If the video is playing when `percySnapshot` is called,
we're going to capture different states of the video. In these cases the
custoemr should provide their own `poster`, which will result in the most stable
screenshots anyways.

* 🏗 Fix karma warning for serving non-js files

* 💚 Wait for video ready state & use webm video

Turns out firefox on ubuntu can't play mpeg videos due to licensing. Swapping to
a webm video makes the tests pass in CI

> Note: Firefox doesn't include MPEG decoding (it's patented), so if that isn't
already in your distribution, you might need to install additional software. Not
sure if there are useful links here or whether a Linux user will need to jump
in: Fix common audio and video issues.

https://support.mozilla.org/en-US/questions/1227605#answer-1138231

* ♻️ Peer review feedback refactoring

* 📝 Update README