♻ Improve build failure handling
When a build fails, the error is logged and the internal queues closed. When attempting to push or
upload new snapshots, the queue throws a `Closed` error. The `sdk-utils` package checks if any API
error have the message `Closed` to disable the local SDK. These `Closed` errors can occasionally
still show up in the logged output which is confusing without any other context (context is usually
further up in the logs anyway).

This handling has now been updated to better recommunicate the original error. When builds fail, the
reason is cached so other methods can rethrow it rather than let the closed error bubble. Our tests
no longer exercised the closed error, so it was made to silently fail for race condition
protection. The CLI API was updated to include build details with errors, and the `sdk-utils`
package updated to check for the build errors rather than a `Closed` error message.