✨ Allow canceling multiple snapshots with alternate syntaxes
Cancelable percy methods were refactored to be represented as generators but return promises by
default. To work directly with the underlying generator, original methods are bound to a new
`percy.yield` property.

The new snapshot syntaxes are made cancelable with this approach so that the normal snapshot method
can remain synchronous. With the old promise/generator method, snaphots would not start until the
returned promise was awaited on. But now the new `percy.yield.snapshot` can be used in generator
contexts (which can be canceled via `generator.throw`).