🐛 Fix supressed validation error causing options to not be scrubbed (#853)
* 🐛 Fix supressed validation error causing option to not be scrubbed

For `allOf`, `anyOf`, and `oneOf`, errors are surpressed since the real error will typically bubble
from one of the nested schemas. However, if those nested schemas are simple primative types, there
is no additional error. And without an error the validation logic does not scrub the option.

This change fixes the issue by adding explicit errors for two of these schemas. In addition, the
execute option was relocated so additional snapshots are not allowed to use an object (only a single
execute function is allowed for additional snapshots).