🐛 Make config normalize skip option behave as expected
Because the option was skipping all or nothing, it would completely skip normalizing higher level
keys if a lower level key was intended to be skipped. By moving the check to within the key map
loop, it allows the oppurtunity to skip any one specific key within a key-path array.