♻ Refactor core generator utils and usage
The generatePromise function now accepts an abort signal and node-style callback. Rather than
returning a promise-like generator, it now always returns a promise.

A new AbortController polyfill was added for Node environments to support providing the
generatorPromise function with an abort signal.

A yieldFor function replaces the previous waitFor function with the same basic functionality;
yielding on every iteration to the provided predicate function.

The waitFor function is now a shortcut for generatePromise(yieldFor(...args)).

A new waitForTimeout function was added in preperation of more exec utils.