🐛 Request fonts as binary via Node (#921)
* 🐛 Request fonts as binary via Node

Sometimes, through the SDKs, fonts are captured as corrupted. This has been a
long standing problem in Percy's SDKs. It turns out we cannot trust the decoding
the browser does in request interception with fonts -- a lot of the time those
fonts network responses are already corrupted by the time our SDK wants to save
the response body.

In order to work around this, we use Node to request the font files directly as
binary data. This solves all issues we've seen with fonts being captured as
corrupt.

* ✅ Add test for new `buffer` `request` option

* ✅ Add test for infering mimetype when it is `text/plain`

* ♻️ PR feedback to handle `false` mime.lookup case & add test