✨ Capture different pixel densities in asset discovery
With the addition of real native devices in our rendering infrastructure, we now
need to capture 2x++ pixel images for pages to render properly (since the real
devices have high pixel density screens).

With the new `deviceScaleFactor` config option, we now will resize the browser
to each specified width at 1x and also whatever the passed value is for
`deviceScaleFactor`. We also run both `beforeResize` and `afterResize` events
with the additional scale resizes.