✨ Add option to delay uploads without deferring until the end
This new `delayUploads` option is an alternative to the existing `deferUploads` option. With the
existing `deferUploads` option, all uploads (including build creation) are deferred until the
instance is stopped. With the new `delayUploads` option, an upload is not processed until the next
upload is queued. So the build is not created until the first snapshot is taken, and that snapshot
is not uploaded until the next snapshot is taken, and so on until the last snapshot is uploaded when
the instance is stopped and the build is finalized.

Currently, this is only useful to speed up the snapshot and storybook commands since those are the
only SDKs that will error when no snapshots are found. In the future, this could become an opt-in
feature for the exec command for the rest of our SDKs, which could error when missing snapshots or
allow retrying queued snapshots.