🐛 Fix static server base-url issue (#1042)
The base-url option for static servers wasn't working correctly for two
reasons. First, is the assumed sitemap URL was being constructed in a way that
would leave off trailing paths from URLs, like how relative links work. Second,
when creating a static server, the generated sitemap was only being served at
the server root, not at the base-url root.

The first issue is fixed by ensuring that any provided base-url has a trailing
forward slash so when the URL is constructed, it treats the base-url as the path
root. The second issue is fixed by adjusting static server creation to host the
generated sitemap at the correct path under the base-url.